requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.10
astroid==2.4.2
attrs==21.2.0
autopep8==1.5.5
boto3==1.15.14
botocore==1.18.14
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==8.0.1
colorama==0.4.4
comtele-sdk==1.1.5
concurrent-videocapture==0.0.2
configparser2==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.6.0
Deprecated==1.2.12
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.45.0
django-appconf==1.0.4
django-avatar==5.0.0
django-base64field==1.0
django-bootstrap-modal-forms==2.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-extra-fields==3.0.2
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-localflavor==3.0.1
django-multiupload==0.5.2
django-oauth-toolkit==1.5.0
django-object-actions==3.0.1
django-pagarme==0.25
django-phonenumber-field==5.0.0
django-pwa==1.0.10
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.0.7
django-s3direct==1.1.8
django-simple-history==2.12.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docusign-esign==3.7.1
dynaconf==3.1.4
et-xmlfile==1.0.1
filelock==3.0.12
filetype==1.0.7
flake8==3.9.2
future==0.18.2
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
googlemaps==4.4.5
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.1
hvac==0.11.0
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==0.9
knox==0.1.14
lazy-object-proxy==1.4.3
loguru==0.5.3
luxand==0.7
MarkupSafe==2.0.1
matrix==2.0.1
mccabe==0.6.1
nose==1.3.7
numpy==1.20.1
oauth2-provider==0.0
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
pagarme-python==4.0.1
phonenumberslite==8.12.23
Pillow==8.0.1
poetry-semver==0.1.0
protobuf==3.17.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
pygame==2.0.1
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python-stdnum==1.15
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
pyzoom==0.1.8
reportlab==3.5.56
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
selenium==3.141.0
shortuuid==1.0.1
simplejson==3.17.3
six==1.15.0
skjold==0.3.2
SQLAlchemy==1.3.23
sqlparse==0.3.1
tinycss2==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.10
validators==0.18.2
virtualenv==20.3.1
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==5.2.0
win32-setctime==1.0.3
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1